+++ ChangeLog.md
@@ -7,6 +7,9 @@
 - Added the `site_description` configuration option. When set, it is used as
   a fallback description for any `head` metadata that uses a description of
   the page. ([#146](https://github.com/davep/blogmore/pull/146))
+- Added the `site_keywords` configuration option. When set, it is used as a
+  a fallback set of keywords for any `head` metadata that uses them.
+  ([#150](https://github.com/davep/blogmore/pull/150))
 
 ## v0.10.0